+++ example/lib/detailscreen.dart
@@ -13,7 +13,6 @@ class _DetailState extends State<Detail> {
 
   @override
   void initState() {
-    // TODO: implement initState
     super.initState();
     WidgetsBinding.instance.addPostFrameCallback(
             (_) {