+++ CHANGELOG.md
@@ -1,7 +1,9 @@
+## [0.1.4] - Added onFinish method [#17](https://github.com/simformsolutions/flutter_showcaseview/issues/17).
+
 ## [0.1.3] - Added feature
 
-# Updated syntax to pass new context to ShowCaseWidget
-# Added onTargetTap callback feature [#10](https://github.com/simformsolutions/flutter_showcaseview/issues/10).
+Updated syntax to pass new context to ShowCaseWidget
+Added onTargetTap callback feature [#10](https://github.com/simformsolutions/flutter_showcaseview/issues/10).
 
 ## [0.1.2] - Fixed issue [#6](https://github.com/simformsolutions/flutter_showcaseview/issues/6).